### 2020-04-30 #37 @apdavison
10 year challenge: Davison, Feng and Brown (2003)
**Original article:** https://doi.org/10.1152/jn.00623.2002

**Scientific domain:**  Computational Neuroscience
**Programming language:**  Hoc
**Suggested editor:**  Tiziano Zito or Georgios Detorakis or Etienne Roesch

Note that this is paper #5 in https://github.com/ReScience/ten-years/issues/1
