### 2020-05-04 #37 @rougier
Thank your for your submission, we'll assign an editor soon.

@otizonaizit @gdetor @eroesch Can you edit this submission for the Ten Years Reproducibility Challenge (only 1 reviewer needed. Reproduction using Neuron and hoc language.
