### 2020-05-19 #37 @otizonaizit
@pdebuyl : would you be up to review this?
