### 2020-05-26 #37 @otizonaizit
@heplesser : do you think you could review this? It is for the [Ten Years Reproducibility Challenge](http://rescience.github.io/ten-years/), with a simplified review process (only one reviewer).
