### 2020-05-26 #37 @otizonaizit
@heplesser : also, have a look at the [guidelines for authors](https://github.com/ReScience/ten-years/blob/master/author-guidelines.md): I did not find them in the website...
