### 2020-05-27 #37 @rougier
I'll review it. See also https://github.com/apdavison/bulbnet-reproduction/issues/1
