### 2020-05-27 apdavison/bulbnet-reproduction#1 @rougier
Some dependencies problem
I'm trying to run the code from the README but I got:
```
python run.py figure_2
NEURON -- VERSION 7.8.0-127-g91f6e623+ master (91f6e623+) 2020-05-08
Duke, Yale, and the BlueBrain Project -- Copyright 1984-2019
See http://neuron.yale.edu/neuron/credits

	1
	1
	1
Reading channel tables
	1
	0
	1001
	0
	0
/Users/rougier/opt/anaconda3/bin/nrniv: table_tabninf_kfasttab undefined function
 in tabchannels.hoc near line 24
 table_tabninf_kfasttab(&datavec[0].x[0], vvec.size, &vvec.x[0])
                                                                ^
        table_tabninf_kfasttab(..., 1001, ...)
      xopen("tabchannels.hoc")
"../../code/ddi_baseline.gnu" line 30: warning: Cannot find or open file "ddi_baseline.curvs"
"../../code/ddi_baseline.gnu" line 30: No data in plot

convert: profile 'icc': 'RGB ': RGB color space not permitted on grayscale PNG `/Users/rougier/tmp/bulbnet-reproduction/results/20200527_192024/ddi_baseline.png' @ warning/png.c/MagickPNGWarningHandler/1748.
```

Any ideas ? A missing dependency maybe ?
