### 2020-05-27 bf2b807
Forgot to mention nrnivmodl in the README
