+++ README.md
@@ -9,8 +9,12 @@ Davison A.P., Feng J. and Brown D. (2003) Dendrodendritic
 inhibition and simulated odor responses in a detailed
 olfactory bulb network model. J. Neurophysiol. 90:1921-35
 
-To reproduce the figures from this paper, change to the code
-directory then run:
+To reproduce the figures from this paper, install NEURON, change to the code
+directory, run:
+
+  $ nrnivmodl
+
+then for each figure run:
 
   $ python run.py figure_2
 
