### 2020-05-27 apdavison/bulbnet-reproduction#1 @apdavison
Probably you need to run "nrnivmodl". I forgot to add this to the README.
