### 2020-05-27 apdavison/bulbnet-reproduction#1 @rougier
Yes, it is now working. If you modify the README, you can also add NEURON and gnuplot as dependencies.
