### 2020-06-09 #37 @rougier
Thanks for the reminder. I'll try to do it by Friday night (I've started playing with the code). Ping me again if not done.
