### 2020-06-12 #37 @rougier
This paper was a real pleasure to review. It is well written and I had no problem to re-run most of the simulations (some were very long and I did no test them) and got the expected results with some discrepancies.

What is a bit troubling though is that for some figures, I got results very similar to the original article and a bit different from the replication (see figure 4B below: I think my re-run is closer to the original). For some other figures, my re-run is a bit farther from both the original and the replication. See for example figure 2b below.  I guess this is linked with the initialization of the
pseudo-random number generator but the qualitative behavior seems a bit impacted. Any idea what could be the explanation ?

Apart from that, I've only a few minor comments and I spotted a typo.

- "Compartmental models of the mitral and granule cells of the had..."
  -> "Compartmental models of the mitral and granule cells of olfactory
      bulb the had..."

- The time estimation of the different simulation could be added in the README since there is no progress indicator during a simulation and you have no idea if the simulation is running as expected.

- In the dependencies, maybe you can add gnuplot as well.

- I was lucky enough to have a warning about a failed conversion of ICC profile that indicated me where was the result. In the output of each simulation, there is a bunch of text but nothing said about where is the actual result.

Overall, the reproduction is really nice, especially given the complexity of the model and the number of results. Congratulations. And I was happy to learn that Gnuplot still exists and seem to be pretty stable.

![bulbnet_schoppa_CNQXAPV](https://user-images.githubusercontent.com/327203/84489282-bdaa7b00-aca1-11ea-9aef-1dd9246b7ec3.png)

![ddi_baseline](https://user-images.githubusercontent.com/327203/84489305-c4d18900-aca1-11ea-8ea5-a7b3fdc01eb0.png)
