### 2020-06-23 #37 @otizonaizit
@apdavison : could you comment on @rougier 's review? I think we are very near to publishing the paper. It would be nice to hear your comments so that we can move forward with this :-)
