### 2020-07-23 fcb5c66
Fixed typo; added more information to README
