+++ reproducibility_challenge_2020/content.tex
@@ -22,7 +22,7 @@ surpassed in both these dimensions, of course (see \cite{Yu:2013}, for example).
 The model consisted of compartmental models of mitral and granule cells, connected by phenomenological models of AMPA, NMDA and \GABAA\ synapses.
 In computational neuroscience, compartmental modelling represents the spatially extended structure of neurons as branching electrical cables,
  then spatially discretizes the cable equation into isopotential compartments \cite{Rall:1967}.
-Compartmental models of the mitral and granule cells of the had been developed previously by \citet{Bhalla:1993} for the GENESIS simulator\citep{GENESIS}, and I intended to build on this basis.
+Compartmental models of the mitral and granule cells of the olfactory bulb had been developed previously by \citet{Bhalla:1993} for the GENESIS simulator\citep{GENESIS}, and I intended to build on this basis.
 However, although I was able to compile GENESIS itself on the DEC Alpha workstation which was the main computational workhorse of the lab, I was unsuccessful in compiling its associated graphical user interface (XODUS),
 probably due to my very limited computational training (self-taught BASIC as a child, then an undergraduate class in FORTRAN for physicists and a graduate course in image processing using MATLAB).
 I therefore decided to attempt to port the models to different simulation software, NEURON \cite{Hines:1997},
