### 2020-07-23 20af571
Added more information to stdout, in particular the location of output files
