### 2020-07-24 f53ae01
Started working on new figure to show random variability
