### 2020-07-27 dbd79dd
Added extension panel to ddi_baseline figure showing variability in IPSC with different random seeds
