### 2020-07-29 2bde26e
Using Sumatra package to insert figures in PDF
