+++ reproducibility_challenge_2020/article.tex
@@ -1,4 +1,5 @@
 \documentclass{rescience}
+\usepackage[recordstore=https://labnotebook.andrewdavison.info/records/, project=bulbnet_reproduction]{sumatra}
 
 \begin{document}
 
