+++ reproducibility_challenge_2020/content.tex
@@ -199,19 +199,22 @@ and similarly for the other figures (2-5 and 7-17; Figure 1 was an illustration 
 For none of the figures were we able to exactly reproduce the orignally published ones, due to differences in the sequences of random numbers.
 %\note{go more into this. We set seeds, I think, but in how many places are RNs used, and are they all governed by the seed. Could try to install older version of NEURON, see if that helps}.
 However, all figures, with the exception of Figure 13 (Figure 11 in this article) gave qualitatively the same result, and were quantitatively very close.
+Figure 2F shows that the result from the original figure lies within the range of results obtained with different random number generator seeds.
 While it is disappointing that we couldn't obtain quantitatively identical results,
 the robustness of the scientific conclusions to differences in random number sequences is a very positive property of the model.
 
 % Figures 2 & 3
 \begin{figure}
     \begin{minipage}{0.48\textwidth}
-        \includegraphics[width=0.96\textwidth]{/home/andrew/projects/bulbnet_reproduction/March2020/results/20200415_151023/ddi_baseline.pdf}
+        \smtincludegraphics[width=0.96\textwidth]{20200727-134735?ddi_random_variability.pdf}
         \caption{Reproduction of Figure 2. The precise number and timing of spikes in the example neuron is different,
                  due to the random element in connecting the networks,
-                 but the shape of the IPSC and the spatial and temporal profiles of the population activity are very similar.}
+                 but the shape of the IPSC and the spatial and temporal profiles of the population activity are very similar.
+                 Panel \textbf{F} is an addition: it shows the IPSC for 50 simulations differing only in the seed to the
+                 random number generator (grey), overlaid with the IPSC from the original Figure 2.}
     \end{minipage}\hfill
     \begin{minipage}{0.48\textwidth}
-        \includegraphics[width=0.96\textwidth]{/home/andrew/projects/bulbnet_reproduction/March2020/results/20200415_151737/ddi_mg.pdf}
+        \smtincludegraphics[width=0.96\textwidth]{20200727-150157?ddi_mg.pdf}
         \caption{Reproduction of Figure 3. The results are visually identical to the original.}
     \end{minipage}
 \end{figure}
@@ -219,14 +222,14 @@ the robustness of the scientific conclusions to differences in random number seq
 % Figures 4 & 5
 \begin{figure}
     \begin{minipage}{0.48\textwidth}
-        \includegraphics[width=0.96\textwidth]{/home/andrew/projects/bulbnet_reproduction/March2020/results/20200430_221540/bulbnet_schoppa_CNQXAPV.pdf}
+        \smtincludegraphics[width=0.96\textwidth]{20200727-151117?bulbnet_schoppa_CNQXAPV.pdf}
         \caption{Reproduction of Figure 4.
         Due to the different, random, connectivity in the reproduction, the individual granule cells show differences
         in their responses (A, B); however at the population level (D), and even at the level of the di-synaptic mitral cell
         response (C), the results are very close to the original, with differences in the fine details of the noise.}
     \end{minipage}\hfill
     \begin{minipage}{0.48\textwidth}
-        \includegraphics[width=0.96\textwidth]{/home/andrew/projects/bulbnet_reproduction/March2020/results/20200415_153039/ddi_nsyn.pdf}
+        \smtincludegraphics[width=0.96\textwidth]{20200727-151518?ddi_nsyn.pdf}
         \caption{Reproduction of Figure 5. %\note{some amplitude differences to look into}
[…]
