+++ reproducibility_challenge_2020/Makefile
@@ -27,7 +27,7 @@ metadata.tex: metadata.yaml
 # missing file reference and interactively asking you for an alternative.
 
 article.pdf: article.tex content.tex header.tex bibliography.bib metadata.tex rescience.cls
-	latexmk -pdf -pdflatex="xelatex -interaction=nonstopmode" -use-make article.tex
+	latexmk -pdf -pdflatex="xelatex -interaction=nonstopmode" -use-make -shell-escape article.tex
 
 clean:
 	@latexmk -CA
