### 2020-08-03 ebb49ef
Updated PDF, with fix to Figure 15 (repr. of original Fig 17)
