+++ reproducibility_challenge_2020/content.tex
@@ -311,7 +311,7 @@ the robustness of the scientific conclusions to differences in random number seq
         \caption{Reproduction of Figure 16. The same comments apply as in the caption to the previous figure.}
     \end{minipage}\hfill
     \begin{minipage}{0.48\textwidth}
-        \smtincludegraphics[width=0.96\textwidth]{20200727-200042?odour_simil.pdf}
+        \smtincludegraphics[width=0.96\textwidth]{20200803-112721?odour_simil.pdf}
         \caption{Reproduction of Figure 17. The assignment of odorant features (`olfactory receptors') to olfactory bulb glomeruli
         (and hence to mitral cells) is random in the model, and hence the input pattern is different in the reproduction compared to
         the original. However, the scientific conclusion of the original, that a small subset of neurons/glomeruli
