### 2020-08-03 #37 @apdavison
Many thanks @rougier for your kind comments.

I've fixed the typo, added information about dependencies and the expected simulation times to the README, and the scripts now print the location of the outputs.

The differences in the figures are certainly due to the random number generator initialisation. For individual granule cell responses (Fig 2B using the numbering from the original manuscript, Fig 1B in the reproduction), the qualitative behaviour can indeed change considerably, as the seed affects the connectivity and hence the number of inputs each cell receives. At the network level, however, the qualitative behaviour is unchanged, as reflected in the mitral cell IPSC (Fig 2C; a disynaptic response) and in the granule cell population spike time histogram (Fig 2E). To confirm this, I've added an extra panel 'F' to Figure 1 (the reproduction of the original Fig. 2) which overlays the original mitral cell IPSC trace from 2003 on top of 50 runs of the simulation with different random seeds. The original trace lies within the range of responses seen with different seeds.

![image](https://user-images.githubusercontent.com/260552/89172206-8a3deb80-d582-11ea-8748-c036c2c93802.png)

By the way, on re-running all of the simulations in a fresh install, I actually get results closer to what you posted above than to the figure in the first version of the manuscript. It is possible I was experimenting with different seeds when I produced the first version of the figures.

Finally, to increase the reproducibility further, I've built the manuscript using the [Sumatra LaTeX package](https://sumatra.readthedocs.io/en/latest/publishing.html): each figure now links to a record on my website of the simulation that produced that figure.
