### 2020-08-04 #37 @rougier
@apdavison Thanks for the explanation. Maybe you can add a small paragraphs in the manuscript with the above explanations.
@otizonaizit I think we can accept the paper. There is one pending suggestion that @apdavison is free to take into account or not but apart from that, I think the paper's ready for publication.
