### 2020-08-04 #37 @otizonaizit
Thanks @rougier !
@apdavison : let me know if you want to add the mentioned paragraph. You'll get a PR from me on your repo to update the metadata before publication.
