### 2020-08-06 f11aa7e
Added more explanation of result variability, as requested by reviewer @rougier
