+++ reproducibility_challenge_2020/metadata.tex
@@ -5,10 +5,10 @@
 \def \codeSWH{}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
+\def \editorNAME{Tiziano Zito}
 \def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
+\def \reviewerINAME{Nicolas P. Rougier}
+\def \reviewerIORCID{0000-0002-6972-589X}
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
 \def \dateRECEIVED{}
@@ -19,7 +19,7 @@
 \def \articleDOMAIN{Computational Neuroscience}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2020}
-\def \reviewURL{}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/37}
 \def \articleABSTRACT{abstract goes here.}
 \def \replicationCITE{Davison A.P., Feng J. and Brown D. (2003) Dendrodendritic inhibition and simulated odor responses in a detailed olfactory bulb network model. Journal of Neurophysiology 90: 1921–1935}
 \def \replicationBIB{dendrodendritic-inhibition-simulated-odor-responses-detailed-olfactory-bulb-network-model}
@@ -29,10 +29,10 @@
 \def \contactEMAIL{andrew.davison@cnrs.fr}
 \def \articleKEYWORDS{rescience c, NEURON simulator, Hoc}
 \def \journalNAME{ReScience C}
-\def \journalVOLUME{4}
+\def \journalVOLUME{6}
 \def \journalISSUE{1}
-\def \articleNUMBER{}
-\def \articleDOI{}
+\def \articleNUMBER{14}
+\def \articleDOI{10.5281/zenodo.3972130}
 \def \authorsFULL{Andrew P. Davison}
 \def \authorsABBRV{A.P. Davison}
 \def \authorsSHORT{Davison}
