+++ reproducibility_challenge_2020/metadata.tex
@@ -11,16 +11,16 @@
 \def \reviewerIORCID{0000-0002-6972-589X}
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
-\def \dateRECEIVED{}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \dateRECEIVED{30 April 2020}
+\def \dateACCEPTED{04 August 2020}
+\def \datePUBLISHED{05 August 2020}
 \def \articleTITLE{[Rp] Dendrodendritic inhibition and simulated odor responses in a detailed olfactory bulb network model}
 \def \articleTYPE{Reproduction}
 \def \articleDOMAIN{Computational Neuroscience}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2020}
 \def \reviewURL{https://github.com/ReScience/submissions/issues/37}
-\def \articleABSTRACT{abstract goes here.}
+\def \articleABSTRACT{This article reports on the experience of reproducing a computational modelling study I originally carried out as part of my PhD studies, which was published as Davison, Feng, and Brown (2003). The reproduced figures were not quantitatively identical to the originals due to differences in the sequences of random numbers obtained, but population responses were quantitatively very similar, and the major findings of the original study are unchanged in the reproduction. We attribute the relative ease of the reproduction to the use of the NEURON simulator and the Gnuplot graphics application, both of which have been continuously maintained since the original publication, and have largely preserved backwards compatibility.}
 \def \replicationCITE{Davison A.P., Feng J. and Brown D. (2003) Dendrodendritic inhibition and simulated odor responses in a detailed olfactory bulb network model. Journal of Neurophysiology 90: 1921–1935}
 \def \replicationBIB{dendrodendritic-inhibition-simulated-odor-responses-detailed-olfactory-bulb-network-model}
 \def \replicationURL{https://journals.physiology.org/doi/pdf/10.1152/jn.00623.2002}
@@ -39,4 +39,4 @@
 \title{\articleTITLE}
 \date{}
 \author[1,\orcid{0000-0002-4793-7541}]{Andrew P. Davison}
-\affil[1]{Institut des Neurosciences Paris-Saclay, UMR 9197 CNRS/Université Paris-Saclay, Gif-sur-Yvette, France}
+\affil[1]{Université Paris-Saclay, CNRS, Institut des neurosciences Paris-Saclay, 91190, Gif-sur-Yvette, France}
