+++ reproducibility_challenge_2020/content.tex
@@ -106,7 +106,7 @@ The code on ModelDB has no explicit licence, in common with the majority of Mode
 This general lack of licenses does not appear to inhibit researchers in computational neuroscience from
 making use of code in ModelDB, nor from publishing new studies based on ModelDB code.
 Since every model in ModelDB is associated with a published article or book chapter,
-the reuse of code is acknowleged primarily through academic citations.
+the reuse of code is acknowledged primarily through academic citations.
 
 % Please provide only information you are certain about. It's OK to forget technical details after many years.
 
@@ -196,7 +196,7 @@ and similarly for the other figures (2-5 and 7-17; Figure~1 was an illustration 
 
 % How close were the results you got to the originally published ones? Include the replicated data table and figure, as we do in other ReScience articles.
 
-For none of the figures were we able to exactly reproduce the orignally published ones, due to differences in the sequences of random numbers.
+For none of the figures were we able to exactly reproduce the originally published ones, due to differences in the sequences of random numbers.
 %\note{go more into this. We set seeds, I think, but in how many places are RNs used, and are they all governed by the seed. Could try to install older version of NEURON, see if that helps}.
 However, all figures, with the exception of Figure~13 (Figure~\ref{fig:13} in this article) gave qualitatively the same result, and were quantitatively very close.
 Responses of individual cells, particularly granule cells (Figure~\ref{fig:2}B) which receive monosynaptic excitation, can be very different from the original,
@@ -362,5 +362,5 @@ It would also have been valuable to go through the source code and systematicall
 were described in the published article; most of them were, but some are missing.
 
 The choice of using well-maintained, stable, widely-used software was a very good one.
-NEURON required no code modifications (although code could be added to take advantage of multi-threading in the newer verions)
+NEURON required no code modifications (although code could be added to take advantage of multi-threading in the newer versions)
 and Gnuplot scripts required minimal modification.
