### 2020-08-07 6e473d0
Added SWH identifier
