+++ reproducibility_challenge_2020/metadata.tex
@@ -2,7 +2,7 @@
 
 \def \codeURL{https://github.com/apdavison/bulbnet-reproduction}
 \def \codeDOI{}
-\def \codeSWH{}
+\def \codeSWH{swh:1:rev:0899bc8af36b50d8cdf5d062562d2721f4602703}
 \def \dataURL{}
 \def \dataDOI{}
 \def \editorNAME{Tiziano Zito}
