+++ reproducibility_challenge_2020/content.tex
@@ -118,7 +118,7 @@ the reuse of code is acknowledged primarily through academic citations.
 As noted above, a version of the source code that reproduced two figures was available on ModelDB
 (it is perhaps impressive that ModelDB still exists, as many neuroinformatics resources funded by the
 US National Institutes of Health (NIH) `Human Brain Project'\footnote{Not to be confused with the current EU research project of the same name}
-in the 1990s are no longer available online.
+in the 1990s are no longer available online.)
 This version of the code was not in itself sufficient to reproduce the other thirteen figures.
 
 The original source code from my thesis, and from additional simulations requested by reviewers during the article publication process,
