### 2020-08-07 #37 @apdavison
Many thanks @rougier, I've added more explanation to the manuscript as you suggested.

@otizonaizit I think everything is now ready to go.
