### 2020-08-10 #37 @otizonaizit
Hi all,

as soon as this PR https://github.com/ReScience/articles/pull/14 and this PR https://github.com/ReScience/rescience.github.io/pull/93 are merged, the paper should appear on the ReScience website!

Congratulations @apdavison !
