### 2020-08-17 #37 @otizonaizit
The paper is online: https://rescience.github.io/read/
